requirements.txt
fastapi
uvicorn[standard]
python-dotenv
gunicorn

langchain
langchain-core
langchain-google-genai
langchain-community
langchain-chroma
langchain-huggingface 
langchain-text-splitters

Pillow
python-multipart
unstructured
markdown

huggingface-hub